Move about window to `about.rs`
